DHT11_FINAL PROJECT/requirements.txt
flask
flask_cors
pyserial
pandas
scikit-learn
joblib
datetime
json
os
time
threading
re
subprocess
